% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_cloud.R
\name{cloud_read}
\alias{cloud_read}
\title{Cloud read}
\usage{
cloud_read(file, bucket = "pulse_cloud")
}
\arguments{
\item{file}{File name of the file to read from Google Cloud Storage. Should
be a csv, xlsx or Rdata.}

\item{bucket}{String containing the GCS-bucket where to read the data from.}
}
\value{
Data.table containing the data of the file specified, or a list with
data.tables.
}
\description{
Download a csv, xlsx or Rdata file from Google Cloud Storage.
}
